reduced boilerplate for surface classes
based on suggestions in issue#7 BenGolberg1++, but less drastic.

Surface is kept as a wrapper class. subs have been converted to
methods and moved from Surface to the existing cairo_surface_t
CPointer class.

End result is reduced boiler-plate, with cairo_surface_t as a raw
class and Surface as a wrapper class.

Shouldn't be any impact to existing users of this module.